client :: moved projects component in the domains folder
