+++ client/src/components/Sidebar/Sidebar.tsx
@@ -1,6 +1,6 @@
 import { useCallback, useState } from "react";
 import { useHistory } from "react-router-dom";
-import { ProjectModal } from "../../pages/Project/ProjectModal";
+import { ProjectModal } from "../../domains/projects/components/ProjectModal";
 import { useToasts } from "../../provider/toastProvider";
 import { Project } from "../../types";
 import { Deferred } from "../../utils/Deferred";
